+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ### API
 * **[BREAKING]** Add `empty()` function to persistent interfaces.
     It allows to get an empty version of a collection without the need to have the related factory.
- 
+* Add factory functions for immutable (non-persistent) collection types
+
 ### Core
 * Add `filterNot` operator